backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.1
pydantic==2.5.3
pydantic[email]
python-dotenv==1.0.0
httpx==0.25.2
gotrue==2.0.0
fastapi[all]
python-jose[cryptography] 